requirements.txt
torch>=2.1.0
pytorch-lightning>=2.1.0
transformers>=4.40.0
torchmetrics>=1.3.0
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0
tqdm>=4.65.0
wandb>=0.15.0
pyyaml>=6.0